requirements.txt
playwright>=1.42.0
playwright-stealth>=1.0.6
pydantic>=2.6.3
pyyaml>=6.0.1
google-generativeai>=0.4.0
aiohttp>=3.9.3
aiosqlite>=0.20.0
tenacity>=8.2.3
pytest>=8.0.0
pytest-asyncio>=0.23.0
aioresponses>=0.7.6
